% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moose_function.R
\name{moose_function}
\alias{moose_function}
\title{A moose function}
\usage{
moose_function(sigma.sq, tau.sq, phi)
}
\arguments{
\item{sigma.sq}{specify the variance}

\item{tau.sq}{specify the measurement error}

\item{phi}{specify the range}
}
\description{
This function allows you to simulate moose harvest data for Northeastern Minnesota.
}
\examples{
cat_function()
}
\keyword{moose}
\keyword{spatial}
